requirements-lisa.txt
# Core ML Libraries
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
matplotlib>=3.7.0
seaborn>=0.12.0

# ML Frameworks
xgboost>=2.0.0
lightgbm>=4.0.0

# Deep Learning (optional)
# torch>=2.0.0  # Uncomment if needed

# Experiment Tracking
mlflow>=2.8.0

# Hyperparameter Optimization
optuna>=3.4.0

# Data Handling
pyyaml>=6.0
python-dateutil>=2.8.0

# Visualization
plotly>=5.17.0

# Progress bars and utilities
tqdm>=4.66.0

# Statistical analysis
scipy>=1.11.0
statsmodels>=0.14.0

# Data validation
# great-expectations>=0.18.0  # Uncomment if needed

# Feature engineering
# featuretools>=1.28.0  # Uncomment if needed
